add shell completion
